requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic>=0.96.0,<1.0
aiosqlite==0.20.0
sse-starlette==2.1.3
slowapi==0.1.9
python-slugify==8.0.4
httpx==0.28.1
python-dotenv==1.0.1
